fix(F-008): split shared-utils.js into 6 cohesive content-script files (behavior-identical)
