+++ apps/caramel-extension/.size-limit.json
@@ -2,8 +2,14 @@
     {
         "name": "content-scripts (injected)",
         "path": [
+            "coupon-constants.generated.js",
             "cart-signals.js",
-            "shared-utils.js",
+            "caramel-base.js",
+            "dom-utils.js",
+            "store-detect.js",
+            "coupon-apply.js",
+            "coupon-fetch.js",
+            "coupon-runner.js",
             "UI-helpers.js",
             "inject.js"
         ],
