+++ apps/caramel-extension/index.html
@@ -59,7 +59,12 @@
 
         <!-- Scripts -->
         <script src="coupon-constants.generated.js"></script>
-        <script src="shared-utils.js"></script>
+        <script src="caramel-base.js"></script>
+        <script src="dom-utils.js"></script>
+        <script src="store-detect.js"></script>
+        <script src="coupon-apply.js"></script>
+        <script src="coupon-fetch.js"></script>
+        <script src="coupon-runner.js"></script>
         <script src="UI-helpers.js"></script>
         <script src="popup.js"></script>
     </body>
