+++ apps/caramel-extension/manifest-firefox.json
@@ -48,7 +48,12 @@
             ],
             "js": [
                 "coupon-constants.generated.js",
-                "shared-utils.js",
+                "caramel-base.js",
+                "dom-utils.js",
+                "store-detect.js",
+                "coupon-apply.js",
+                "coupon-fetch.js",
+                "coupon-runner.js",
                 "UI-helpers.js",
                 "inject.js",
                 "amazon.js"
