+++ apps/caramel-extension/manifest.json
@@ -39,7 +39,12 @@
             "js": [
                 "coupon-constants.generated.js",
                 "cart-signals.js",
-                "shared-utils.js",
+                "caramel-base.js",
+                "dom-utils.js",
+                "store-detect.js",
+                "coupon-apply.js",
+                "coupon-fetch.js",
+                "coupon-runner.js",
                 "UI-helpers.js",
                 "inject.js"
             ]
