+++ apps/caramel-extension/scripts/test-extension.mjs
@@ -9,8 +9,12 @@
  *   4. Supported store sample has valid selectors
  *   5. Coupons endpoint reachable
  *   6. Popup UI: fills the login form and verifies logged-in state
- *   7. Injection logic: loads shared-utils.js against a fake store DOM and
- *      verifies applyCoupon() finds the input, fills the code, and clicks apply.
+ *   7. Injection logic: loads the F-008 split content-script files (formerly
+ *      one shared-utils.js — coupon-constants.generated.js, caramel-base.js,
+ *      dom-utils.js, store-detect.js, coupon-apply.js, coupon-fetch.js,
+ *      coupon-runner.js, in real manifest.json load order) against a fake
+ *      store DOM and verifies applyCoupon() finds the input, fills the
+ *      code, and clicks apply.
  *
  * Prereqs:
  *   - caramel-app dev server running on localhost:58000 (pnpm dev)
@@ -198,9 +202,21 @@ async function main() {
 
         // 8. DOM INJECTION — real applyCoupon() against a synthetic supported-store DOM
         {
-            const sharedUtilsSrc = readFileSync(
-                path.join(EXT_PATH, 'shared-utils.js'),
-                'utf8',
+            // F-008 split shared-utils.js into 6 files; coupon-fetch.js also
+            // needs coupon-constants.generated.js loaded first (F-006's
+            // RESTRICTED_STATUSES rebind reads window.CaramelCoupons at
+            // module-eval time). Real manifest.json/index.html load order.
+            const contentScriptFiles = [
+                'coupon-constants.generated.js',
+                'caramel-base.js',
+                'dom-utils.js',
+                'store-detect.js',
+                'coupon-apply.js',
+                'coupon-fetch.js',
+                'coupon-runner.js',
+            ]
+            const contentScriptSources = contentScriptFiles.map(f =>
+                readFileSync(path.join(EXT_PATH, f), 'utf8'),
             )
 
             const page = await context.newPage()
@@ -212,8 +228,12 @@ async function main() {
                 <button id="apply-btn">Apply</button>
             </body></html>`)
 
-            // Wire click handler and load shared-utils.js via evaluate (bypasses CSP)
-            await page.evaluate(utilsSrc => {
+            // Wire click handler and load the split content-script files, IN
+            // ORDER, via evaluate (bypasses CSP) — each its own eval, like
+            // separate <script> tags, so load-order semantics match the
+            // real content-script realm (see caramel-base.js's relocated
+            // _isDevInstall for why that distinction matters).
+            await page.evaluate(sources => {
                 window.__clickLog = []
                 document
                     .getElementById('apply-btn')
@@ -223,7 +243,7 @@ async function main() {
                         })
                         document.getElementById('total').textContent = '$90.00'
                     })
-                // Stub chrome APIs that shared-utils.js touches at module scope
+                // Stub chrome APIs the content-script files touch at module scope
                 // Force-overwrite: Chromium provides `chrome` on about:blank but without .runtime
                 window.chrome = {
                     runtime: {
@@ -236,8 +256,10 @@ async function main() {
                     },
                     storage: { sync: { get: (_, cb) => cb && cb({}) } },
                 }
-                ;(0, eval)(utilsSrc)
-            }, sharedUtilsSrc)
+                for (const src of sources) {
+                    ;(0, eval)(src)
+                }
+            }, contentScriptSources)
 
             const rec = {
                 domain: 'test-store.local',
