+++ apps/caramel-extension/tests/popup.test.mjs
@@ -1,15 +1,17 @@
 import { beforeAll, describe, expect, it } from 'vitest'
-import { loadExtensionSource } from './_load.mjs'
+import { loadExtensionSource, loadExtensionSources } from './_load.mjs'
 
 // F-002 UI pin — proves the honest-failure plumbing all the way through:
 // background.js now replies {error:'HTTP <status>'} on a non-ok upstream
-// fetch (background.test.mjs); shared-utils.js's fetchCoupons() already
-// throws on resp.error; popup.js's initPopup() already catches that and
-// renders the load-error state instead of silently falling through to
-// "no coupons for this site" (which would misrepresent an OUTAGE as a
-// factual absence of coupons — the bug this finding exists for).
+// fetch (background.test.mjs); coupon-fetch.js's fetchCoupons() (formerly
+// shared-utils.js, split by F-008 — move-only, cat-diff-proven
+// behavior-identical) already throws on resp.error; popup.js's
+// initPopup() already catches that and renders the load-error state
+// instead of silently falling through to "no coupons for this site"
+// (which would misrepresent an OUTAGE as a factual absence of coupons —
+// the bug this finding exists for).
 //
-// Goes through the real listener chain (shared-utils.js's fetchCoupons,
+// Goes through the real listener chain (coupon-fetch.js's fetchCoupons,
 // unstubbed) rather than re-implementing the shaping — only the messaging
 // transport (currentBrowser.runtime.sendMessage/storage.sync.get) is
 // stubbed, since there's no real background service worker in this harness.
@@ -20,25 +22,38 @@ beforeAll(() => {
         '<div id="loading-container"></div><div id="auth-container"></div>'
 
     // F-006 — coupon-constants.generated.js sets window.CaramelCoupons,
-    // which shared-utils.js now reads unconditionally at module-eval time
+    // which coupon-fetch.js now reads unconditionally at module-eval time
     // (RESTRICTED_STATUSES's rebind) and popup.js reads when rendering a
     // coupon list. Real load order (manifest.json, manifest-firefox.json,
     // index.html) always puts it first; mirror that here.
     loadExtensionSource('coupon-constants.generated.js', [])
 
-    // Load shared-utils.js next: it establishes window/globalThis
-    // .currentBrowser (via the chrome stub installed for THIS call) and
-    // defines the real fetchCoupons(). popup.js references both as free
-    // globals (see its `/* global currentBrowser, fetchCoupons */` header)
-    // and does not declare either itself.
-    loadExtensionSource('shared-utils.js', [])
+    // Load the 6 F-008 split files next, in real manifest load order and
+    // sharing ONE chrome stub (loadExtensionSources): together they
+    // establish window/globalThis.currentBrowser (caramel-base.js's
+    // bootstrap) and define the real fetchCoupons() (coupon-fetch.js).
+    // popup.js references both as free globals (see its
+    // `/* global currentBrowser, fetchCoupons */` header) and does not
+    // declare either itself.
+    loadExtensionSources(
+        [
+            'caramel-base.js',
+            'dom-utils.js',
+            'store-detect.js',
+            'coupon-apply.js',
+            'coupon-fetch.js',
+            'coupon-runner.js',
+        ],
+        [],
+    )
 
     // Stub the messaging transport on the stub captured above. A later
     // loadExtensionSource() call (for popup.js, below) installs its OWN
     // fresh chrome stub and reassigns globalThis.chrome — but that does not
     // affect this object, since currentBrowser already holds a direct
-    // reference to it (captured at shared-utils.js load time, not a live
-    // re-read of globalThis.chrome).
+    // reference to it (captured at caramel-base.js's bootstrap, the first
+    // of the loadExtensionSources() call above, not a live re-read of
+    // globalThis.chrome).
     globalThis.currentBrowser.runtime.sendMessage = (message, cb) => {
         if (message?.action === 'getActiveTabDomainRecord') {
             cb({ url: 'https://example.com/cart' })
